# docs/swagger.yaml
basePath: /
definitions:
...
  ttlock.EKeyChangeValidTimeRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      endDate:
        type: integer
      keyId:
        type: integer
      startDate:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - endDate
    - keyId
    - startDate
    type: object
  ttlock.EKeyDeleteRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      keyId:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - keyId
    type: object
  ttlock.EKeyFreezeUnfreezeRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      keyId:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - keyId
    type: object
  ttlock.EKeyGetOneRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      lockId:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - lockId
    type: object
  ttlock.EKeyListRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      groupId:
        type: integer
      lockAlias:
        type: string
      pageNo:
        type: integer
      pageSize:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - pageNo
    - pageSize
    type: object
  ttlock.EKeySendRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      createUser:
        enum:
        - 1
        - 2
        type: integer
      date:
        type: integer
      endDate:
        type: integer
      keyName:
        type: string
      keyRight:
        enum:
        - 1
        - 2
        type: integer
      lockId:
        type: integer
      receiverUsername:
        type: string
      remarks:
        type: string
      remoteEnable:
        enum:
        - 1
        - 2
        type: integer
      startDate:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - endDate
    - keyName
    - lockId
    - receiverUsername
    - startDate
    type: object
  ttlock.EKeyUpdateRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      keyId:
        type: integer
      keyName:
        type: string
      remoteEnable:
        enum:
        - 1
        - 2
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - keyId
    type: object
...
info:
  contact: {}
  title: TTLock Api
  version: "1.0"
paths:
...
  /api/ekey/change-valid-time:
    post:
      consumes:
      - application/json
      description: Change eKey validity period
      parameters:
      - description: eKey valid time request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.EKeyChangeValidTimeRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: eKey valid time changed successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Change eKey valid time
      tags:
      - EKeys
  /api/ekey/delete:
    post:
      consumes:
      - application/json
      description: Delete an eKey
      parameters:
      - description: eKey delete request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.EKeyDeleteRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: eKey deleted successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Delete eKey
      tags:
      - EKeys
  /api/ekey/freeze:
    post:
      consumes:
      - application/json
      description: Freeze an eKey
      parameters:
      - description: eKey freeze request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.EKeyFreezeUnfreezeRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: eKey frozen successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Freeze eKey
      tags:
      - EKeys
  /api/ekey/get-one:
    post:
      consumes:
      - application/json
      description: Get eKey details by ID
      parameters:
      - description: eKey get one request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.EKeyGetOneRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: eKey fetched successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get one eKey
      tags:
      - EKeys
  /api/ekey/list:
    post:
      consumes:
      - application/json
      description: Get all eKeys
      parameters:
      - description: eKey list request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.EKeyListRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: eKeys fetched successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get eKey list
      tags:
      - EKeys
  /api/ekey/send:
    post:
      consumes:
      - application/json
      description: Send an eKey to a user
      parameters:
      - description: eKey send request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.EKeySendRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: eKey sent successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Send eKey
      tags:
      - EKeys
  /api/ekey/unfreeze:
    post:
      consumes:
      - application/json
      description: Unfreeze an eKey
      parameters:
      - description: eKey unfreeze request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.EKeyFreezeUnfreezeRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: eKey unfrozen successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Unfreeze eKey
      tags:
      - EKeys
  /api/ekey/update:
    post:
      consumes:
      - application/json
      description: Update eKey information
      parameters:
      - description: eKey update request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.EKeyUpdateRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: eKey updated successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Update eKey
      tags:
      - EKeys
...
swagger: "2.0"